inline JimpleValueSwitch, change retValue to option
